[REFACTOR] Temp 파일들 원복
