+++ app/src/main/java/org/sopt/santamanitto/main/MainViewModel.kt
@@ -8,9 +8,10 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import org.sopt.santamanitto.NetworkViewModel
-import org.sopt.santamanitto.room.data.TempMyManittoModel
+import org.sopt.santamanitto.room.data.MyManittoModel
 import org.sopt.santamanitto.room.network.RoomRequest
 import org.sopt.santamanitto.util.RoomSortUtil
+import timber.log.Timber
 import javax.inject.Inject
 
 @HiltViewModel
@@ -18,8 +19,8 @@ class MainViewModel @Inject constructor(
     private val roomRequest: RoomRequest
 ) : NetworkViewModel() {
 
-    private var _myManittoModelList = MutableSharedFlow<List<TempMyManittoModel>>(replay = 1)
-    val myManittoModelList: SharedFlow<List<TempMyManittoModel>> = _myManittoModelList
+    private var _myManittoModelList = MutableSharedFlow<List<MyManittoModel>>(replay = 1)
+    val myManittoModelList: SharedFlow<List<MyManittoModel>> = _myManittoModelList
 
     private val _isRefreshing = MutableStateFlow(false)
     val isRefreshing: StateFlow<Boolean> = _isRefreshing
@@ -30,10 +31,13 @@ class MainViewModel @Inject constructor(
             _isLoading.value = true
             try {
                 val rooms = roomRequest.getRooms()
+                Timber.e("rooms: $rooms")
                 _myManittoModelList.emit(RoomSortUtil.getSortedRooms(rooms))
             } catch (e: Exception) {
+                Timber.e(e)
                 _networkErrorOccur.value = true
             } finally {
+                Timber.e("finally")
                 _isLoading.value = false
                 _isRefreshing.value = false
             }
@@ -64,6 +68,8 @@ class MainViewModel @Inject constructor(
     }
 
     fun refresh() {
+        Timber.e("refresh1")
         fetchMyManittoList()
+        Timber.e("refresh2")
     }
 }
