+++ app/src/main/java/org/sopt/santamanitto/main/list/RemovedMyManittoViewHolder.kt
@@ -3,13 +3,13 @@ package org.sopt.santamanitto.main.list
 import android.view.ViewGroup
 import org.sopt.santamanitto.R
 import org.sopt.santamanitto.databinding.ItemMymanittoRemovedBinding
-import org.sopt.santamanitto.room.data.TempMyManittoModel
+import org.sopt.santamanitto.room.data.MyManittoModel
 import org.sopt.santamanitto.view.recyclerview.BaseViewHolder
 
 class RemovedMyManittoViewHolder(
     parent: ViewGroup,
     removeListener: ((roomId: String) -> Unit)?
-) : BaseViewHolder<TempMyManittoModel, ItemMymanittoRemovedBinding>(
+) : BaseViewHolder<MyManittoModel, ItemMymanittoRemovedBinding>(
     R.layout.item_mymanitto_removed,
     parent
 ) {
@@ -26,7 +26,7 @@ class RemovedMyManittoViewHolder(
         }
     }
 
-    override fun bind(data: TempMyManittoModel) {
+    override fun bind(data: MyManittoModel) {
         roomId = data.roomId
     }
