+++ app/src/main/java/org/sopt/santamanitto/room/data/PersonalRoomModel.kt
@@ -1,11 +1,10 @@
 package org.sopt.santamanitto.room.data
 
 import com.google.gson.annotations.SerializedName
+import org.sopt.santamanitto.room.data.MyManittoModel.Member.Manitto
+import org.sopt.santamanitto.room.data.MyManittoModel.Mission
 
-// TODO: 추후 삭제
 data class PersonalRoomModel(
-    @SerializedName("SantaUserId") val santaUserId: String,
-    @SerializedName("ManittoUserId") val manittoUserId: String,
-    @SerializedName("MyMission") val myMission: MissionContentModel?,
-    @SerializedName("MissionToMe") val missionToMe: MissionContentModel?
+    @SerializedName("manitto") val manitto: Manitto,
+    @SerializedName("mission") val mission: Mission
 )
