+++ app/src/main/java/org/sopt/santamanitto/room/network/RoomRequest.kt
@@ -3,8 +3,9 @@ package org.sopt.santamanitto.room.network
 import org.sopt.santamanitto.room.create.network.CreateRoomModel
 import org.sopt.santamanitto.room.create.network.CreateRoomRequestModel
 import org.sopt.santamanitto.room.create.network.ModifyRoomRequestModel
-import org.sopt.santamanitto.room.data.TempMyManittoModel
-import org.sopt.santamanitto.room.data.TempPersonalRoomModel
+import org.sopt.santamanitto.room.data.MyManittoModel
+import org.sopt.santamanitto.room.data.PersonalRoomModel
+import org.sopt.santamanitto.room.join.network.JoinRoomModel
 import org.sopt.santamanitto.room.join.network.JoinRoomRequestModel
 import org.sopt.santamanitto.room.join.network.JoinRoomResponseModel
 import org.sopt.santamanitto.room.manittoroom.network.ManittoRoomModel
@@ -30,7 +31,7 @@ interface RoomRequest {
     }
 
     interface GetPersonalRoomInfoCallback {
-        fun onLoadPersonalRoomInfo(personalRoom: TempPersonalRoomModel)
+        fun onLoadPersonalRoomInfo(personalRoom: PersonalRoomModel)
 
         fun onDataNotAvailable()
     }
@@ -39,7 +40,7 @@ interface RoomRequest {
         WrongInvitationCode, AlreadyMatched, AlreadyEntered, Els
     }
 
-    suspend fun getRooms(): List<TempMyManittoModel>
+    suspend fun getRooms(): List<MyManittoModel>
 
     fun createRoom(request: CreateRoomRequestModel, callback: CreateRoomCallback)
